IMPROV: Change order of travis run
Changed the order of the travis scripts to avoid format errors after
waiting for the static analysis to complete.
Moved back to markdownlint-cli with a controlled node environment and
a fixed tool version, to avoid errors as in the past.
Skipped Markdown and Python check during travis run when only one
commit is pushed to the pull request.

Related #303